Backport fixes for RUSTSEC-2026-0119
This vulnerability appears fixed in just a few commits, but due to the drift
between 0.25 and 0.26, a larger amount of commits had to be backported to
reconcile the two. Thankfully, all of these commits together did not have any
merge conflicts and seems to be _mostly_ API compatible. A few functions were
trimmed between 0.25 and 0.26, and that's reflected here.

Specifically, this is a squash merge of the following commits:
- 2865fd53b24177ca8ddc372bea079db483a4047e Remove no-op BinEncoder method
- 13260f20c4d0c2f955b352b990bb5ad7f0d57e31 Remove unused method
- e90e4bed62fd96a2172bdadcfc7bc9282fb9ad0f Replace canonical_names flag w/ name encoding enum
- dce701fec9fe6659c92af86905265fbd782df171 Add setter and getter for name encoding mode
- f9acbc9954a68d57766549d1414baf81564e8a16 Add canonical form flag, set when encoding TBS
- ab547a2c2479bea495ba27872dac638a282c7e17 Use name encoding mode directly
- 6bd7c29e8791867d88d74fe7f81cfda26cec7e0c Add ModalEncoder
- d153bdd17c8a6fddc2fd75bcae0fc0cdcba05086 Add method to switch name encoding mode for RDATA
- f5848c06a2ceb6a5339f8d9387e039cb706d4876 Set RDATA name encoding mode based on record type
- e2d24657db3f3993a86d7ef7a0583272de64e534 Remove Name::emit_with_lowercase()
- 333ea203c57fb3704c3e940b2cad30f3e3cdc179 Remove Name::emit_as_canonical()
- 6e70e931515286a701b00ce91b06988eb71e89ef Limit number of name pointers stored in BinEncoder
- a9b884b27223937d9a9afa933d7859d5b0784b61 Skip search when name compression is disabled
- f09321075b1f97902b7bc4ca4ffda7816fcf2971 Limit number of names that may be compressed

All commits are contained to `hickory-proto`.
`cargo check --all-features --all-targets` and
`cargo test -p hickory-proto --all-features` pass as expected, except for a
QUIC test which has a certificate which expired 220 days ago. Despite these
basic checks, this degree of cherry-picking by someone without familiarity with
the project (kayabaNerve) can still be considered a non-trivial risk towards
the integrity of the result. Unfortunately, for those who can't update to
`hickory-proto 0.26` (such as when it's a transient dependency), this may still
be preferable.